split up dashboard to many components + add one time balance update on  reg
