+++ src/components/DashBoard/Dashboard.module.css
@@ -15,326 +15,18 @@
   margin-bottom: 2rem;
 }
 
-.card {
-  background-color: white;
-  padding: 1.5rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.statsCard {
-  margin-bottom: 5px;
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.iconWrapper {
-  padding: 0.75rem;
-  border-radius: 9999px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.iconWrapperBlue {
-  background-color: #dbeafe;
-}
-
-.iconWrapperRed {
-  background-color: #fee2e2;
-}
-
-.iconWrapperGreen {
-  background-color: #dcfce7;
-}
-
-.iconWrapperPurple {
-  background-color: rgba(147, 51, 234, 0.1);
-}
-
-.iconWrapperTeal {
-  background-color: rgba(20, 184, 166, 0.1);
-}
-
-.icon {
-  width: 1.5rem;
-  height: 1.5rem;
-}
-
-.iconBlue {
-  color: #2563eb;
-}
-
-.iconRed {
-  color: #dc2626;
-}
-
-.iconGreen {
-  color: #16a34a;
-}
-
-.iconPurple {
-  color: rgb(147, 51, 234);
-}
-
-.iconTeal {
-  color: rgb(20, 184, 166);
-}
-
-.statsContent {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.statsLabel {
-  color: #6b7280;
-  font-size: 0.875rem;
-}
-
-.statsValue {
-  font-size: 1.5rem;
-  font-weight: 700;
-  line-height: 1;
-}
-
 .title {
   font-size: 1.125rem;
   font-weight: 600;
   margin-bottom: 0.5rem;
 }
 
-.chartContainer {
-  width: 100%;
-  height: 300px;
-  overflow: hidden;
-}
-
-.chartHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.chartControls {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.timeRangeButtons {
-  display: flex;
-  gap: 10px;
-}
-
-.timeRangeButton {
-  padding: 8px 16px;
-  border: 1px solid #e2e8f0;
-  border-radius: 6px;
-  background: white;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.timeRangeButton:hover {
-  background: #f1f5f9;
-}
-
-.timeRangeButtonActive {
-  background: #3b82f6;
-  color: white;
-  border-color: #3b82f6;
-}
-
-.timeRangeButtonActive:hover {
-  background: #2563eb;
-}
-
-.tooltip {
-  background-color: white;
-  border-radius: 0.375rem;
-  padding: 1rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  border: 1px solid #e5e7eb;
-}
-
-.tooltipText {
-  font-size: 0.875rem;
-  font-weight: 600;
-  margin-bottom: 0.5rem;
-}
-
-.tooltipEntry {
-  font-size: 0.875rem;
-  margin-bottom: 0.25rem;
-}
-
-.input {
-  width: 100%;
-  padding: 0.5rem;
-  border: 1px solid #d1d5db;
-  border-radius: 0.375rem;
-  margin-bottom: 0.5rem;
-}
-
-.button {
-  width: 100%;
-  background-color: #3b82f6;
-  color: white;
-  padding: 0.5rem;
-  border-radius: 0.375rem;
-  text-align: center;
+.setupNameLink {
+  color: #007bff;
+  text-decoration: underline;
   cursor: pointer;
 }
 
-.button:hover {
-  background-color: #2563eb;
-}
-
-.button:disabled {
-  background-color: #93c5fd;
-  cursor: not-allowed;
-}
-
-.valueRed {
-  color: #ef4444;
-}
-
-.valueGreen {
-  color: #10b981;
-}
-
-.forecastSection {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 1rem;
-  margin-bottom: 1rem;
-}
-
-.goalForecast {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.probability {
-  font-size: 0.875rem;
-  color: #6b7280;
-}
-
-.forecastToggle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.875rem;
-  color: #6b7280;
-  cursor: pointer;
-}
-
-.forecastToggle input[type="checkbox"] {
-  width: 1rem;
-  height: 1rem;
-  cursor: pointer;
-}
-
-.forecastMetrics {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  margin-top: 0.5rem;
-  padding-top: 0.5rem;
-  border-top: 1px solid #e5e7eb;
-}
-
-.metricLabel {
-  font-size: 0.75rem;
-  color: #6b7280;
-}
-
-.metricValue {
-  font-size: 1rem;
-  font-weight: 600;
-}
-
-.predictionAccuracy {
-  display: inline-flex;
-  align-items: center;
-  padding: 0.25rem 0.5rem;
-  border-radius: 9999px;
-  font-size: 0.75rem;
-  font-weight: 500;
-}
-
-.accuracyHigh {
-  background-color: #dcfce7;
-  color: #16a34a;
-}
-
-.accuracyMedium {
-  background-color: #fef9c3;
-  color: #854d0e;
-}
-
-.accuracyLow {
-  background-color: #fee2e2;
-  color: #dc2626;
-}
-
-.trendIndicator {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-  font-size: 0.875rem;
-}
-
-.trendUp {
-  color: #16a34a;
-}
-
-.trendDown {
-  color: #dc2626;
-}
-
-.forecastChart {
-  position: relative;
-  margin-top: 1rem;
-}
-
-.chartLegend {
-  display: flex;
-  gap: 1rem;
-  margin-top: 0.5rem;
-  justify-content: center;
-}
-
-.legendItem {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-  font-size: 0.75rem;
-  color: #6b7280;
-}
-
-.legendDot {
-  width: 0.5rem;
-  height: 0.5rem;
-  border-radius: 9999px;
-}
-
-.noData {
-  text-align: center;
-  padding: 2rem;
-  color: #6b7280;
-  font-style: italic;
-}
-.forecastToggleDisabled {
-  color: #888;
-  cursor: not-allowed;
-}
-
-.forecastToggleDisabled input {
-  cursor: not-allowed;
-  opacity: 0.5;
+.setupNameLink:hover {
+  color: #0056b3;
 }
